Refactor tabcomplete to be in commands
