+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -205,9 +205,9 @@ import at.hannibal2.skyhanni.features.misc.items.EstimatedWardrobePrice
 import at.hannibal2.skyhanni.features.misc.items.GlowingDroppedItems
 import at.hannibal2.skyhanni.features.misc.massconfiguration.DefaultConfigFeatures
 import at.hannibal2.skyhanni.features.mining.powdertracker.PowderTracker
-import at.hannibal2.skyhanni.features.misc.tabcomplete.GetFromSacksTabComplete
-import at.hannibal2.skyhanni.features.misc.tabcomplete.PlayerTabComplete
-import at.hannibal2.skyhanni.features.misc.tabcomplete.WarpTabComplete
+import at.hannibal2.skyhanni.features.commands.tabcomplete.GetFromSacksTabComplete
+import at.hannibal2.skyhanni.features.commands.tabcomplete.PlayerTabComplete
+import at.hannibal2.skyhanni.features.commands.tabcomplete.WarpTabComplete
 import at.hannibal2.skyhanni.features.misc.teleportpad.TeleportPadCompactName
 import at.hannibal2.skyhanni.features.misc.teleportpad.TeleportPadInventoryNumber
 import at.hannibal2.skyhanni.features.misc.tiarelay.TiaRelayHelper
