+++ src/main/java/at/hannibal2/skyhanni/features/commands/tabcomplete/WarpTabComplete.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.features.misc.tabcomplete
+package at.hannibal2.skyhanni.features.commands.tabcomplete
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
